+++ kbe/src/lib/common/rsa.cpp
@@ -217,7 +217,7 @@ std::string KBE_RSA::encrypt(const std::string& instr)
 
 	char strencrypted[1024];
 	memset(strencrypted, 0, 1024);
-	strutil::bytes2string((unsigned char *)encrypted.data(), encrypted.size(), (unsigned char *)strencrypted, 1024);
+	strutil::bytes2string((unsigned char *)encrypted.data(), static_cast<int>(encrypted.size()), (unsigned char *)strencrypted, 1024);
 	return strencrypted;
 }
 
@@ -228,7 +228,7 @@ int KBE_RSA::encrypt(const std::string& instr, std::string& outCertifdata)
 
 	unsigned char* certifdata =(unsigned char*)calloc(RSA_size(static_cast<RSA*>(rsa_public)) + 1, sizeof(unsigned char));
 
-	int certifsize = RSA_public_encrypt(instr.size(),
+	int certifsize = RSA_public_encrypt(static_cast<int>(instr.size()),
 		(unsigned char*)instr.c_str(), certifdata, static_cast<RSA*>(rsa_public), RSA_PKCS1_OAEP_PADDING);
 
 	if (certifsize < 0)
