+++ kbe/src/lib/entitydef/datatypes.cpp
@@ -11,7 +11,7 @@ DataTypes::DATATYPE_MAP DataTypes::dataTypesLowerName_;
 DataTypes::UID_DATATYPE_MAP DataTypes::uid_dataTypes_;
 DataTypes::DATATYPE_ORDERS DataTypes::dataTypesOrders_;
 
-static uint8 _g_baseTypeEndIndex = 0;
+static size_t _g_baseTypeEndIndex = 0;
 
 //-------------------------------------------------------------------------------------
 DataTypes::DataTypes()
