+++ kbe/src/lib/pyscript/scriptstderr.cpp
@@ -91,7 +91,7 @@ PyObject* ScriptStdErr::__py_write(PyObject* self, PyObject *args)
 		return NULL;
 	}
 		
-	static_cast<ScriptStdErr*>(self)->pScriptStdOutErr()->error_msg(sdata, size);
+	static_cast<ScriptStdErr*>(self)->pScriptStdOutErr()->error_msg(sdata, static_cast<uint32>(size));
 	S_Return;
 }
