+++ kbe/src/server/tools/kbcmd/client_sdk_plugin.cpp
@@ -78,8 +78,8 @@ bool ClientSDKPlugin::create(const std::string& path)
 	);
 
 
-	int written = fwrite(sourcefileBody_.c_str(), 1, sourcefileBody_.size(), fp);
-	if (written != (int)sourcefileBody_.size())
+	size_t written = fwrite(sourcefileBody_.c_str(), 1, sourcefileBody_.size(), fp);
+	if (written != sourcefileBody_.size())
 	{
 		ERROR_MSG(fmt::format("ClientSDK::saveFile(): fwrite error! {}\n",
 			configPath));
